AboutUsInfo reorganised
