+++ src/App.css
@@ -1,23 +1,12 @@
 
 :root {
-  --primary-color: #eaeaea;
+  --main-color: #eaeaea;
   --main-font: "Gothic A1", sans-serif;
-  --sm-margin: 8px;
-  --md-margin: 16px;
-  --lg-margin: 32px;
-  --max-margin: 48px;
-  --xs-font: 16px;
-  --sm-font: 18px;
-  --md-font: 20px;
-  --lg-font: 24px;
-  --base-padding: 16px;
+  --standard-margin: 16px;
+  --small-margin: 8px;
+  --main-font: 16px;
+  --large-font: 24px;
+  --standard-padding: 16px;
   --standard-shadow: 2px 2px 4px rgba(0, 0, 0, 0.29);
 }
 
-.about-us {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: 0 var(--base-padding);
-  justify-content: center;
-}
