+++ src/components/AboutUsInfo/AboutUsInfo.module.css
@@ -1,0 +1,7 @@
+.aboutUs {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    padding: 0 var(--standard-padding);
+    justify-content: center;
+  }
