+++ src/pages/LandingPage.js
@@ -1,10 +1,7 @@
 import React from "react";
-import { AboutUsInfo } from "../components/AboutUsInfo";
+import { AboutUsInfo } from "../components/AboutUsInfo/AboutUsInfo";
+import {RandomText} from "../components/AboutUsInfo/RandomText"
 
-const RandomText ={
-  title: 'About us',
-  text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
-}
 
 function LandingPage() {
     return(
